+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 jui extension Change Log
 2.0.8 under development
 -----------------------
 
-- no changes in this release.
+- Bug #87: Inline display does not allow using altField option (buttflattery)
 
 
 2.0.7 November 25, 2017